apps/authentication/serializers.py
from rest_framework import serializers

class UserSerializer(serializers.Serializer):
    username = serializers.CharField(read_only=True)
    permissions = serializers.ListSerializer(child=serializers.CharField(read_only=True))